refactor creating ProfileSet
