+++ src/lib/profile_set.js
@@ -1,3 +1,16 @@
+export function createProfileSet(profiles, userInfo, settings) {
+  const {
+    loginDisplayNameAccount, loginDisplayNameUser,
+    roleDisplayNameAccount, roleDisplayNameUser,
+  } = userInfo;
+  const { showOnlyMatchingRoles } = settings;
+
+  const baseAccount = brushAccountId(loginDisplayNameAccount);
+  const filterByTargetRole = showOnlyMatchingRoles ? (roleDisplayNameUser || loginDisplayNameUser.split("/", 2)[0]) : null;
+
+  return new ProfileSet(profiles, baseAccount, { filterByTargetRole });
+}
+
 export class ProfileSet {
   constructor(items, baseAccount, { filterByTargetRole }) {
     // Map that has entries { <awsAccountId>: <Profile> }
@@ -48,4 +61,10 @@ export class ProfileSet {
     }
     return profiles;
   }
+}
+
+function brushAccountId(val) {
+  const r = val.match(/^(\d{4})\-(\d{4})\-(\d{4})$/);
+  if (!r) return val;
+  return r[1] + r[2] + r[3];
 }
